#122: Moved duplicate code from FoldLeft and FoldRight to Fold base class. Moved containsEmpty() method into that base class as it is the only user.
